Rename combineFrames file and vars
- fix historical naming to reflect existing naming patterns.
